Added shapes presentation tier

+++ VGA_core_M4/lib/logic_tier/shapes_data.h
@@ -108,7 +108,7 @@ status_t VGA_setPrimitiveData(	sPosition_t* 		centerPointPos,
 								uint16_t 			rotationDegrees,
 								color_t				lineColor,
 								color_t				fillColor,
-								uint16_t			lineWeight,
+								uint8_t				lineWeight,
 								primitiveShape_t	primitiveShape	);
 
 #endif /* SHAPES_DATA_H_ */
